refactor: book, review, image 리팩토링
### 작업 내용
- 책 단건 등록
요청 Body의 카테고리를 필수 입력 값으로 변경하였습니다.

- open api 책 등록
기존 인증된 유저의 닉네임을 통해 검색하는 방법에서 Param을 통해 검색어를 직접 입력하도록 수정하였습니다.
검색 조건을 publisher(출판사)에서 total(전체)로 수정하였습니다.

- 책 & 리뷰 이미지 관련 파일 분리
- 이미지 등록 개수 제한
- 가격 타입 BigDecimal로 변경

+++ src/main/java/com/example/ililbooks/domain/book/dto/response/BookResponse.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Getter;
 import org.springframework.data.domain.Page;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -20,7 +21,7 @@ public class BookResponse {
 
     private final String author;
 
-    private final Long price;
+    private final BigDecimal price;
 
     private final String category;
 
@@ -39,7 +40,7 @@ public class BookResponse {
     private final LocalDateTime modifiedAt;
 
     @Builder
-    private BookResponse(Long id, Long userId, String title, String author, Long price, String category, int stock, Page<ReviewResponse> reviews, String imageUrl, SaleStatus saleStatus, LimitedType limitedType, LocalDateTime createdAt, LocalDateTime modifiedAt) {
+    private BookResponse(Long id, Long userId, String title, String author, BigDecimal price, String category, int stock, Page<ReviewResponse> reviews, String imageUrl, SaleStatus saleStatus, LimitedType limitedType, LocalDateTime createdAt, LocalDateTime modifiedAt) {
         this.id = id;
         this.userId = userId;
         this.title = title;
